requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
distlib==0.3.7
Django==5.0
django-allauth==0.60.1
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-tailwind==3.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.12.4
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-secret-manager==2.16.4
google-cloud-speech==2.21.0
googleapis-common-protos==1.60.0
gorilla-cli==0.0.10
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
gunicorn==21.2.0
halo==0.0.31
idna==3.4
log-symbols==0.0.14
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.9
six==1.16.0
spinners==0.0.24
sqlparse==0.4.4
tee==0.0.3
termcolor==2.3.0
tzdata==2023.3
urllib3==2.0.5
vertexai==0.0.1
virtualenv==20.24.5
wcwidth==0.2.8
